c/requirements.txt
grpcio
grpcio-tools
aiohttp
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-grpc
opentelemetry-instrumentation-aiohttp-client